+++ NEWS.md
@@ -1,3 +1,7 @@
+# crew.cluster 0.3.3 (development)
+
+
+
 # crew.cluster 0.3.3
 
 * Add `slurm_memory_gigabytes_required` to set `--mem` in SLURM (#44, @multimeric).